Move kubefed2/federate/disable to kubefed2 pkg
