+++ pkg/kubefed2/disable.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package federate
+package kubefed2
 
 import (
 	"context"
